*move Test::RegionDetection and Test::DetectionParam to Synet.
Move RegionDetection and DetectionParam from the Test namespace into
Synet/Decoders/Detection.h so the detection facade can be used outside
the test projects.
