+++ prj/vs2022/Synet.vcxproj
@@ -142,6 +142,7 @@
     <ClInclude Include="..\..\src\Synet\Decoders\Alpha.h" />
     <ClInclude Include="..\..\src\Synet\Decoders\Anchor.h" />
     <ClInclude Include="..\..\src\Synet\Decoders\DetOut.h" />
+    <ClInclude Include="..\..\src\Synet\Decoders\Detection.h" />
     <ClInclude Include="..\..\src\Synet\Decoders\Iim.h" />
     <ClInclude Include="..\..\src\Synet\Decoders\Nanodet.h" />
     <ClInclude Include="..\..\src\Synet\Decoders\Region.h" />
