+++ prj/vs2022/Synet.vcxproj.filters
@@ -458,6 +458,9 @@
     <ClInclude Include="..\..\src\Synet\Decoders\DetOut.h">
       <Filter>Decoders</Filter>
     </ClInclude>
+    <ClInclude Include="..\..\src\Synet\Decoders\Detection.h">
+      <Filter>Decoders</Filter>
+    </ClInclude>
     <ClInclude Include="..\..\src\Synet\Decoders\Yolo.h">
       <Filter>Decoders</Filter>
     </ClInclude>
