+++ prj/vs2022/TestInferenceEngine.vcxproj
@@ -37,7 +37,6 @@
     <ClInclude Include="..\..\src\Test\TestOutputComparer.h" />
     <ClInclude Include="..\..\src\Test\TestParams.h" />
     <ClInclude Include="..\..\src\Test\TestPerformance.h" />
-    <ClInclude Include="..\..\src\Test\TestRegionDetection.h" />
     <ClInclude Include="..\..\src\Test\TestReport.h" />
     <ClInclude Include="..\..\src\Test\TestSynet.h" />
     <ClInclude Include="..\..\src\Test\TestUtils.h" />
