+++ prj/vs2022/TestMultiThreads.vcxproj
@@ -41,7 +41,6 @@
     <ClInclude Include="..\..\src\Test\TestOutputComparer.h" />
     <ClInclude Include="..\..\src\Test\TestParams.h" />
     <ClInclude Include="..\..\src\Test\TestPerformance.h" />
-    <ClInclude Include="..\..\src\Test\TestRegionDetection.h" />
     <ClInclude Include="..\..\src\Test\TestReport.h" />
     <ClInclude Include="..\..\src\Test\TestUtils.h" />
   </ItemGroup>
