+++ src/Test/TestSynet.h
@@ -27,7 +27,6 @@
 #include "TestCommon.h"
 #include "TestPerformance.h"
 #include "TestNetwork.h"
-#include "TestRegionDetection.h"
 
 #include "Synet/Network.h"
 
@@ -138,7 +137,7 @@ namespace Test
         virtual Regions GetRegions(const Size & size, float threshold, float overlap) const
         {
             if (_regionDetection.Enable())
-                return _regionDetection.GetRegions(_net, size, threshold, overlap, CTX_NUM - 1);
+                return _regionDetection.GetRegions(_net, size.x, size.y, threshold, overlap, CTX_NUM - 1);
             else
                 return _net.GetRegions(size.x, size.y, threshold, overlap, CTX_NUM - 1);
         }
@@ -154,7 +153,7 @@ namespace Test
         bool _trans, _sort;
         Floats _lower, _upper;
         size_t _synetMemoryUsage;
-        RegionDetection _regionDetection;
+        Synet::RegionDetection _regionDetection;
         const int CTX_NUM = 1;
 
         bool Load(const String & model, const String & weight, const Options& options)
